Exchange: Improve the parser for the Truncated element
Some exchange server configures the Truncated element with a string
value ("true" for TRUE, "false" for FALSE) instead of a boolean
value (1 for TRUE and 0 for FALSE). So it causes a IOException and
then throw a ParserException.

Improve the parser to check the Truncated element value and return
the correct result even if the value is of the string type.

CRs-fixed: 661309

Change-Id: I2d1c810737d0dc27d87ecc259428e90682d3d9f4